# docs/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: manageCallAI API
  version: 0.1.0
  summary: Business-level telecom control plane API over FreeSWITCH
  description: |
    OpenAPI contract for the manageCallAI REST API.

    **MVP response envelope**: all implemented endpoints return `{ "data": ... }` for
    both single-resource and list responses. Pagination will be introduced later
    where required. Error responses return `{ "error": "<CODE>", "message": "<human text>",
    "request_id": "<id>" }` where `error` is a machine-readable RPC code (NOT_FOUND,
    INVALID_ARGUMENT, UNAUTHENTICATED, PERMISSION_DENIED, ALREADY_EXISTS,
    RESOURCE_EXHAUSTED, INTERNAL, UNAVAILABLE).

    **Auth model**:
    - Human-facing endpoints: `Authorization: Bearer <JWT>` (issued by `/auth/login`
      or `/auth/register`). JWT claims include `sub`, `tenant_id`, `email`, `role`.
    - Runtime/internal endpoints: `Authorization: Bearer <RUNTIME_API_TOKEN>` or
      header `x-managecallai-runtime-token: <token>`. These are called by
      FreeSWITCH and the Go ESL agent — not by end users.

    Only implemented endpoints are documented here. Future resources (outbound
    routes, prompts, CDRs, audit events, approvals, simulations) will be added
    when implemented.
servers:
  - url: https://api.example.com/api/v1
    description: Production placeholder
  - url: http://localhost:3000/api/v1
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /queues:
    get:
      tags:
        - Queues
      summary: List queues
      responses:
        "200":
          description: Queue list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueListResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    post:
      tags:
        - Queues
      summary: Create queue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQueueRequest"
      responses:
        "201":
          description: Queue created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /queues/{queueId}:
    parameters:
      - $ref: "#/components/parameters/QueueIdParam"
    get:
      tags:
        - Queues
      summary: Get queue
      responses:
        "200":
          description: Queue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    patch:
      tags:
        - Queues
      summary: Update queue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQueueRequest"
      responses:
        "200":
          description: Queue updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /queues/{queueId}/deactivate:
    parameters:
      - $ref: "#/components/parameters/QueueIdParam"
    post:
      tags:
        - Queues
      summary: Deactivate queue
      responses:
        "200":
          description: Queue deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /queues/{queueId}/members:
    parameters:
      - $ref: "#/components/parameters/QueueIdParam"
    post:
      tags:
        - Queues
      summary: Add queue member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddQueueMemberRequest"
      responses:
        "201":
          description: Queue member created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueMemberResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /queues/{queueId}/members/{extensionId}:
    parameters:
      - $ref: "#/components/parameters/QueueIdParam"
      - $ref: "#/components/parameters/ExtensionIdParam"
    delete:
      tags:
        - Queues
      summary: Remove queue member
      responses:
        "204":
          description: Queue member removed
        default:
          $ref: "#/components/responses/ErrorResponse"
...
  /recordings:
    get:
      tags:
        - Recordings
      summary: List recordings for the authenticated tenant
      description: |
        Lists tenant-owned call and voicemail recording metadata. Public responses
        intentionally expose storage-safe metadata only; raw storage internals and
        temporary media URLs are not part of this contract.
      parameters:
        - name: call_id
          in: query
          schema:
            type: string
            minLength: 1
            maxLength: 255
      responses:
        "200":
          description: Recording list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecordingListResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
  /recordings/{recordingId}:
    get:
      tags:
        - Recordings
      summary: Get a recording for the authenticated tenant
      parameters:
        - $ref: "#/components/parameters/RecordingIdParam"
      responses:
        "200":
          description: Recording
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecordingResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /recordings/{recordingId}/playback:
    get:
      tags:
        - Recordings
      summary: Stream tenant-owned recording media
      description: |
        Streams recording media only when the resolved file is under the configured
        recording storage root. The endpoint is tenant-scoped and does not expose
        raw filesystem paths in its response.
      parameters:
        - $ref: "#/components/parameters/RecordingIdParam"
      responses:
        "200":
          description: Recording media stream
          content:
            audio/wav:
              schema:
                type: string
                format: binary
        default:
          $ref: "#/components/responses/ErrorResponse"
  /recordings/internal/ingest:
    post:
      tags:
        - Recordings
      summary: Ingest completed recording metadata from the runtime agent
      description: |
        Runtime-internal endpoint used by the FreeSWITCH agent or another trusted
        runtime component after a call recording or voicemail recording is available.
      security:
        - runtimeTokenHeader: []
        - runtimeTokenBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngestRecordingRequest"
      responses:
        "201":
          description: Recording metadata created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecordingResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /recordings/{recordingId}/analysis-requests:
    get:
      tags:
        - Recordings
      summary: List analysis requests for a recording
      description: |
        Lists transcription and summarization requests for a tenant-owned recording.
        Results may be produced by an external plugin, worker, or AI service.
      parameters:
        - $ref: "#/components/parameters/RecordingIdParam"
      responses:
        "200":
          description: Recording analysis request list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecordingAnalysisRequestListResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    post:
      tags:
        - Recordings
      summary: Request transcription and/or summary for a recording
      description: |
        Creates a provider-neutral analysis request for a tenant-owned recording.
        `preferred_provider` is only a hint; implementations may ignore it or route
        it to a user-installed plugin. `whisper` is documented as an example provider
        value, not as a required platform dependency.
      parameters:
        - $ref: "#/components/parameters/RecordingIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRecordingAnalysisRequest"
      responses:
        "201":
          description: Analysis request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecordingAnalysisRequestResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /recordings/{recordingId}/analysis-requests/{analysisRequestId}:
    get:
      tags:
        - Recordings
      summary: Get analysis request status and result
      parameters:
        - $ref: "#/components/parameters/RecordingIdParam"
        - $ref: "#/components/parameters/AnalysisRequestIdParam"
      responses:
        "200":
          description: Recording analysis request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecordingAnalysisRequestResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: User JWT issued by /auth/login or /auth/register
    runtimeTokenBearer:
      type: http
      scheme: bearer
      description: RUNTIME_API_TOKEN as Bearer token (preferred)
    runtimeTokenHeader:
      type: apiKey
      in: header
      name: x-managecallai-runtime-token
      description: RUNTIME_API_TOKEN as custom header
  parameters:
    ExtensionIdParam:
      name: extensionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    QueueIdParam:
      name: queueId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    RecordingIdParam:
      name: recordingId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    AnalysisRequestIdParam:
      name: analysisRequestId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          enum:
            - NOT_FOUND
            - INVALID_ARGUMENT
            - UNAUTHENTICATED
            - PERMISSION_DENIED
            - ALREADY_EXISTS
            - RESOURCE_EXHAUSTED
            - INTERNAL
            - UNAVAILABLE
        message:
          type: string
        request_id:
          type: string
      required:
        - error
        - message
        - request_id
...